:label: [#445] Convert ValidationErrors display in submission summary to TS
